requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
PyMySQL==1.1.2
python-dotenv==1.0.0
itsdangerous==2.2.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Werkzeug==3.1.6
WTForms==3.1.2
gunicorn==21.2.0
openpyxl==3.1.2
python-dateutil==2.8.2
